backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.6.1
openai==0.27.10
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
bcrypt==4.1.2
docker==7.0.0
stellar-sdk==13.1.0
ipfshttpclient==0.8.0a2